requirements.txt
asgiref==3.8.1
cffi==1.17.1
colorama==0.4.6
cryptography==44.0.3
Django==5.1
et-xmlfile==1.1.0
mysqlclient==2.2.7
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxlsb==1.0.10
six==1.16.0
sqlparse==0.5.1
tqdm==4.67.1
tzdata==2024.1
